+++ src/menu.rs
@@ -16,10 +16,10 @@ fn print_both(menu: &[&str], prompt: &str) {
 
 
 // Grab the next line of input
-fn next_input() -> Option<uint> {
+fn next_input() -> Option<usize> {
 
     let line = io::stdin().read_line().unwrap();
-    let input: Option<uint> = line.trim().parse();
+    let input: Option<usize> = line.trim().parse();
     input
 }
