+++ src/prime_decomposition.rs
@@ -3,7 +3,7 @@
 use std::num::Float;
 
 // We need this to be public because it is used from another file
-pub fn factor(mut nb: uint) -> Vec<uint> {
+pub fn factor(mut nb: usize) -> Vec<usize> {
     let mut result = vec!();
 
     // First we take out all even factors.
@@ -14,14 +14,14 @@ pub fn factor(mut nb: uint) -> Vec<uint> {
 
     // Then (if any left) we take out the odd ones.
     let mut cand = 3;
-    let mut max_bound = (nb as f32).sqrt() as uint + 1;
+    let mut max_bound = (nb as f32).sqrt() as usize + 1;
 
     while cand <= max_bound {
         while nb % cand == 0 {
             result.push(cand);
             nb /= cand;
         }
-        max_bound = (nb as f32).sqrt() as uint + 1;
+        max_bound = (nb as f32).sqrt() as usize + 1;
         cand += 2;
     }
