+++ src/roots_of_unity.rs
@@ -12,7 +12,7 @@ fn main() {
     }
 }
 
-fn roots_of_unity(degree: uint) -> Vec<Complex32> {
+fn roots_of_unity(degree: usize) -> Vec<Complex32> {
     range(0, degree).map(|el|
         Complex::<f32>::from_polar(&1f32, &(2f32 * consts::PI * (el as f32) / (degree as f32))))
         .collect::<Vec<Complex32>>()
